combined color maps and filter
